Get rid off old Interval file
